Componentize Election UI (#610)
* Break down ongoing-round into subcomponents.
* Add ephemeral chain test to set up and start election.
* Set election time.
