+++ .github/workflows/build.yml
@@ -126,7 +126,7 @@ jobs:
                   cp eden-micro-chain.wasm ../product_cache
                   cp eden.abi ../product_cache
                   cp eden.wasm ../product_cache
-                  cp run-full-election.wasm ../product_cache
+                  cp run-elections.wasm ../product_cache
                   cp run-genesis.wasm ../product_cache
                   cp token.abi ../product_cache
                   cp token.wasm ../product_cache
@@ -181,7 +181,7 @@ jobs:
                       product_cache/token.abi
                       product_cache/token.wasm
                       product_cache/run-genesis.wasm
-                      product_cache/run-full-election.wasm
+                      product_cache/run-elections.wasm
 
     build-micro-chain:
         name: Build Micro Chain
