+++ contracts/eden/tests/run-elections.cpp
@@ -2,7 +2,7 @@
 
 TEST_CASE("Setup Eden chain with full election")
 {
-   nodeos_runner r("chain-full-election");
+   nodeos_runner r("chain-run-elections");
 
    r.tester.genesis();
    r.tester.run_election(true, 10000, true);
@@ -10,6 +10,9 @@ TEST_CASE("Setup Eden chain with full election")
    r.tester.induct_n(100);
    r.checkpoint("inductions");
    r.tester.run_election(true, 10000, true);
+   r.checkpoint("full_election");
+   r.tester.eden_gm.act<actions::electsettime>(s2t("2021-11-18T04:12:00.000"));
+   r.tester.start_election(true, 10000);
 
    r.start_nodeos();
 }
