+++ packages/webapp/src/elections/components/index.ts
@@ -3,5 +3,4 @@ export * from "./election-member-chips";
 export * from "./election-community-room-button";
 export * from "./error-loading-election";
 export * from "./ongoing-election";
-export * from "./pie-mer";
 export * from "./registration-election";
